+++ lib/sentinel/registrator.ex
@@ -39,7 +39,15 @@ defmodule Sentinel.Registrator do
     |> Changeset.put_change(:hashed_password, hashed_password)
   end
   def set_hashed_password(changeset) do
+    if !is_invitable? do
+      changeset = changeset
+      |> Changeset.add_error(:password, "can't be blank")
+    end
+
     changeset
-    |> Changeset.add_error(:password, "can't be blank")
+  end
+
+  defp is_invitable? do
+    Application.get_env(:sentinel, :invitable) || false
   end
 end