+++ test/controller/user_controller_test.exs
@@ -3,9 +3,13 @@ defmodule UserControllerTest do
   use Plug.Test
 
   import RouterHelper
+  alias Sentinel.Registrator
+  alias Sentinel.PasswordResetter
+  alias Sentinel.Confirmator
   alias Sentinel.TestRouter
   alias Sentinel.TestRepo
   alias Sentinel.User
+  alias Sentinel.UserHelper
   alias Mix.Config
   import Sentinel.Util
 
@@ -28,6 +32,7 @@ defmodule UserControllerTest do
 
   test "default sign up" do
     Config.persist([sentinel: [confirmable: :optional]])
+    Config.persist([sentinel: [invitable: false]])
 
     conn = call(TestRouter, :post, "/api/users", %{user: %{password: @password, email: @email}}, @headers)
     assert conn.status == 201
@@ -49,6 +54,7 @@ defmodule UserControllerTest do
 
   test "confirmable :required sign up" do
     Config.persist([sentinel: [confirmable: :required]])
+    Config.persist([sentinel: [invitable: false]])
 
     conn = call(TestRouter, :post, "/api/users", %{user: %{password: @password, email: @email}}, @headers)
     assert conn.status == 201
@@ -67,7 +73,8 @@ defmodule UserControllerTest do
   end
 
   test "confirmable :false sign up" do
-    Config.persist([sentinel: [confirmable: :false]])
+    Config.persist([sentinel: [confirmable: false]])
+    Config.persist([sentinel: [invitable: false]])
 
     conn = call(TestRouter, :post, "/api/users", %{user: %{password: @password, email: @email}}, @headers)
     assert conn.status == 201
@@ -82,6 +89,72 @@ defmodule UserControllerTest do
     assert length(Mailman.TestServer.deliveries) == 0
   end
 
+  test "invitable sign up" do
+    Config.persist([sentinel: [invitable: true]])
+    Config.persist([sentinel: [confirmable: false]])
+
+    conn = call(TestRouter, :post, "/api/users", %{user: %{email: @email}}, @headers)
+    assert conn.status == 201
+    assert Poison.decode!(conn.resp_body) == "ok"
+
+    mail =  Mailman.TestServer.deliveries |> List.last |> Mailman.Email.parse!
+
+    assert mail.from == "test@example.com"
+    assert mail.to == ["User <#{@email}>"]
+    assert mail.subject == "You've been invited to Test App " <> @email
+    assert mail.text == "Hello user@example.com!\n\n\You've been invited to Test App\n\n"
+  end
+
+  test "invitable and confirmable sign up" do
+    Config.persist([sentinel: [confirmable: :optional]])
+    Config.persist([sentinel: [invitable: true]])
+
+    conn = call(TestRouter, :post, "/api/users", %{user: %{email: @email}}, @headers)
+    assert conn.status == 201
+    assert Poison.decode!(conn.resp_body) == "ok"
+
+    mail =  Mailman.TestServer.deliveries |> List.last |> Mailman.Email.parse!
+
+    assert mail.from == "test@example.com"
+    assert mail.to == ["User <#{@email}>"]
+    assert mail.subject == "You've been invited to Test App " <> @email
+    assert mail.text == "Hello user@example.com!\n\n\You've been invited to Test App\n\n"
+  end
+
+  test "invitable setup password" do
+    Config.persist([sentinel: [confirmable: :optional]])
+    Config.persist([sentinel: [invitable: true]])
+
+    {confirmation_token, changeset} = Registrator.changeset(%{email: @email})
+                                      |> Confirmator.confirmation_needed_changeset
+    user = repo.insert!(changeset)
+
+    {password_reset_token, changeset} = PasswordResetter.create_changeset(user)
+    user = repo.update!(changeset)
+
+    conn = call(TestRouter, :post, "/api/users/#{user.id}/invited", %{confirmation_token: confirmation_token, password_reset_token: password_reset_token, password: @password}, @headers)
+    assert conn.status == 201
+    %{"token" => token} = Poison.decode!(conn.resp_body)
+
+    assert repo.one(GuardianDb.Token).jwt == token
+    updated_user = repo.get! UserHelper.model, user.id
+
+    assert updated_user.hashed_confirmation_token == nil
+    assert updated_user.hashed_password_reset_token == nil
+    assert updated_user.unconfirmed_email == nil
+  end
+
+  test "sign up with missing password without the invitable module enabled" do
+    Config.persist([sentinel: [invitable: false]])
+    conn = call(TestRouter, :post, "/api/users", %{user: %{email: @email}}, @headers)
+    assert conn.status == 422
+
+    errors = Poison.decode!(conn.resp_body)
+              |> Dict.fetch!("errors")
+
+    assert errors["password"] == "can't be blank"
+  end
+
   test "sign up with missing email" do
     conn = call(TestRouter, :post, "/api/users", %{"user" => %{"password" => @password}}, @headers)
     assert conn.status == 422
@@ -92,17 +165,10 @@ defmodule UserControllerTest do
     assert errors["email"] == "can't be blank"
   end
 
-  test "sign up with missing password" do
-    conn = call(TestRouter, :post, "/api/users", %{user: %{email: @email}}, @headers)
-    assert conn.status == 422
-
-    errors = Poison.decode!(conn.resp_body)
-              |> Dict.fetch!("errors")
-
-    assert errors["password"] == "can't be blank"
-  end
-
   test "sign up with custom validations" do
+    Config.persist([sentinel: [confirmable: :optional]])
+    Config.persist([sentinel: [invitable: false]])
+
     Application.put_env(:sentinel, :user_model_validator, fn changeset ->
       Ecto.Changeset.add_error(changeset, :password, "too_short")
     end)