+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
-## 0.0.6
+## 0.1.0
 Merges in @termoose's changes that declutter compilation output
 
 Updates password reset with unknown email to always return 200, to
 prevent malicious fishing for user emails
+
+Adds invitable module, accessible by adding `invitible: true` to your
+sentinel config
 
 ## 0.0.5
 Makes `guardian_db` dependency optional